thor/pyproject.toml
[project]
name = "thor"
version = "0.1.0"
description = ""
authors = [
    {name = "theNullP0inter",email = "nikhilallamsetti@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery[redis] (>=5.5.2,<6.0.0)",
    "starlette (>=0.47.0,<0.48.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "asgiref (>=3.8.1,<4.0.0)",
    "typer (>=0.16.0,<0.17.0)",
    "zmq (>=0.0.0,<0.0.1)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "mcp (>=1.9.4,<2.0.0)"
]


[tool.poetry]
packages = [
    {include = "thor", from = "src"}
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
thor = "thor.cli:app"